範囲選択機能の追加
BeatMapEditor.cppに範囲選択用のスプライトを初期化するコードを追加し、選択エリアの描画機能を実装しました。`SelectNote`関数に範囲選択の条件を追加し、`ClearSelection`関数で範囲選択フラグをリセットする処理を追加しました。`HandleInput`関数ではドラッグ操作による範囲選択の処理を実装し、`CheckSelectionArea`関数を新たに追加しました。また、`BeatMapEditor.h`に新しい関数の宣言を追加しました。

+++ Application/Source/Application/BeatMapEditor/LiveMapping/LiveMapping.cpp
@@ -17,7 +17,7 @@ void LiveMapping::Initialize(int32_t _laneCount)
 
 void LiveMapping::Update(float _elapsedTime)
 {
-    for (size_t i = 0; i < laneKeyBindings_.size(); ++i)
+    for (uint32_t i = 0; i < laneKeyBindings_.size(); ++i)
     {
         // キーが押されたかチェック
         if (input_->IsKeyTriggered(laneKeyBindings_[i]))
